backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.6.0
email-validator>=2.1.0.post1
httpx>=0.26.0